+++ ios/Classes/Converters/SQRDCard+FlutterReaderSDKAdditions.m
@@ -65,6 +65,9 @@ limitations under the License.
         case SQRDCardBrandSquareGiftCard:
             result = @"SQUARE_GIFT_CARD";
             break;
+        case SQRDCardBrandEftpos:
+            result = @"EFTPOS";
+            break;
     }
     return result;
 }