+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### v1.1.0 Mar 5, 2019
+
+* remove `alwaysRequireSignature` and add `collectSignature` to checkout configuration.
+* bump the minimum dependency to Reader SDK 1.1.1(iOS)/1.1.3(Android).
+* this change **does NOT** include all new features introduced in Reader SDK 1.1.* such as **Store customer card**, see reader SDK [Change Log](https://docs.connect.squareup.com/changelog/mobile-logs/2019-02-13) for details.
+
 ### v1.0.3 Dec 5, 2018
 
 * Documentation update.